fix(spells): 修复动态常量咒语中的游戏时间类型转换问题
- 将游戏时间从 long 类型转换为 double 类型以匹配预期的数值类型
- 确保咒语返回正确的数值类型避免潜在的类型不匹配错误

+++ src/main/java/org/creepebucket/programmable_magic/spells/spells_compute/DynamicConstantSpell.java
@@ -39,7 +39,7 @@ public abstract class DynamicConstantSpell extends SpellItemLogic implements Spe
 
         @Override
         public ExecutionResult run(Player caster, SpellSequence spellSequence, List<Object> paramsList, SpellEntity spellEntity) {
-            return ExecutionResult.RETURNED(this, List.of(caster.level().getGameTime()), List.of(SpellValueType.NUMBER));
+            return ExecutionResult.RETURNED(this, List.of((double) caster.level().getGameTime()), List.of(SpellValueType.NUMBER));
         }
     }
